requirements.txt
discord.py==1.5.1
praw==7.1.0
pymongo==3.11.2
dnspython==2.0.0
